Unique fingerprints.

+++ src/logging.h
@@ -16,16 +16,16 @@ public:
     void set_parent(Progress* parent) { _parent = parent; }
 
     double progress_stage() { return _cur_progress; }
-    double progress_total() { if (_cur_stage >= _costs.size()) return 1; int cost = 0; for (int i = 0; i < _cur_stage; cost += _costs[i], i++); return (cost + _costs[_cur_stage]*_cur_progress)/_total_cost; }
-    int cost_total() { return _total_cost; }
+    double progress_total() { if (_cur_stage >= _costs.size()) return 1; double cost = 0; for (int i = 0; i < _cur_stage; cost += _costs[i], i++); return (cost + _costs[_cur_stage]*_cur_progress)/_total_cost; }
+    double cost_total() { return _total_cost; }
     double time_total() { return _time_total + _time_stage; }
     double time_op() { return _time_op; }
     int op_count() { return _op_count; }
     int num_stages() { return (int)_costs.size(); }
 
 private:
-    std::vector<int> _costs;
-    int _total_cost = 0;
+    std::vector<double> _costs;
+    double _total_cost = 0;
     int _cur_stage = 0;
     double _cur_progress = 0;
     double _time_total = 0;
